examples/fastapi-echo/requirements.txt
line-bot-sdk
fastapi
uvicorn[standard]
anthropic
